perf(Config): Use caches for config files
Config files are read many times during runtime. In native filesystems,
it is not a big issue. However, when using the app with remote
filesystems, the performance degrades a lot. Even, in some cases, it
crashes. This change check first if the config data is in the cache. If
not, it reads the config file from the filesystem and updates the cache.